Breaking down the Vorepanel.jsx file
