+++ tgui/packages/tgui/interfaces/chompstation/VorePanel/VoreUserPreferenceItem.tsx
@@ -1,0 +1,24 @@
+import { useBackend } from '../../../backend';
+import { Button } from '../../../components';
+
+export const VoreUserPreferenceItem = (props) => {
+  const { act } = useBackend();
+
+  const { spec, ...rest } = props;
+  const { action, test, tooltip, content, fluid = true, back_color } = spec;
+
+  return (
+    <Button
+      onClick={() => act(action)}
+      icon={test ? 'toggle-on' : 'toggle-off'}
+      selected={test}
+      fluid={fluid}
+      backgroundColor={
+        back_color ? (test ? back_color.enabled : back_color.disabled) : ''
+      }
+      tooltip={tooltip.main + ' ' + (test ? tooltip.disable : tooltip.enable)}
+      content={test ? content.enabled : content.disabled}
+      {...rest}
+    />
+  );
+};
