+++ tgui/packages/tgui/interfaces/chompstation/VorePanel/constants.ts
@@ -1,0 +1,45 @@
+export const stats = [null, 'average', 'bad'];
+
+export const digestModeToColor = {
+  Default: null,
+  Hold: null,
+  Digest: 'red',
+  Absorb: 'purple',
+  Unabsorb: 'purple',
+  Drain: 'orange',
+  Selective: 'orange',
+  Shrink: 'teal',
+  Grow: 'teal',
+  'Size Steal': 'teal',
+  Heal: 'green',
+  'Encase In Egg': 'blue',
+};
+
+export const reagentToColor = {
+  Water: null,
+  Milk: null,
+  Cream: null,
+  Honey: 'teal',
+  'Cherry Jelly': 'teal',
+  'Digestive acid': 'red',
+  'Diluted digestive acid': 'red',
+  'Space cleaner': null,
+  Lube: null,
+  Biomass: 'teal',
+  'Concentrated Radium': 'orange',
+  Tricordrazine: 'green',
+};
+
+export const digestModeToPreyMode = {
+  Hold: 'being held.',
+  Digest: 'being digested.',
+  Absorb: 'being absorbed.',
+  Unabsorb: 'being unabsorbed.',
+  Drain: 'being drained.',
+  Selective: 'being processed.',
+  Shrink: 'being shrunken.',
+  Grow: 'being grown.',
+  'Size Steal': 'having your size stolen.',
+  Heal: 'being healed.',
+  'Encase In Egg': 'being encased in an egg.',
+};
